Fix Ticket Creation
I finally fixed the Customer Ticket relations and have relations to it. I might change it back to have what I had before in order to streamline things. I'm not sure.

Local DB had to be deleted and replaced. I'm not sure why, but now it has the correct tables.

ProgressList got changed and now uses a List<SelectListItem> and when selected in CreateTicket creates a string value that is stored in the Ticket table.

Folders also recieved a name change from Plural to Singular.

Details on Customer will also show a list of current tickets for the Customer. Details for Ticket also work and show the appropriate details. Might change later to show more of the Customer's Info (It definitely should).

(´･ω･`)(´･ω･`)